switch to wiki pageviews
