--- tests/tasks/uipath-maestro-flow/stock_price/stock_price.yaml
@@ -1,54 +1,0 @@
-task_id: skill-flow-stock-price
-description: >
-  Create a UiPath Flow that takes a ticker and two dates, fetches the closing
-  price from Yahoo Finance for each, and returns the difference in cents.
-  Exercises managed HTTP (manual auth) called twice in the same flow and
-  deterministic scoring via historical prices.
-tags: [uipath-maestro-flow, e2e, generate, ootb, http]
-max_iterations: 1
-
-agent:
-  type: claude-code
-  turn_timeout: 1200
-
-sandbox:
-  driver: tempdir
-  python: {}
-
-initial_prompt: |
-  Create a UiPath Flow project named "StockPrice" that takes `symbol`
-  (string), `date1`, and `date2` (numbers, unix timestamps in seconds) as
-  inputs. Use Yahoo Finance's chart API
-  (`https://query1.finance.yahoo.com/v8/finance/chart/{symbol}?period1={date}&period2={date+86400}&interval=1d`)
-  to fetch the closing price for each date, then return the difference
-  (date2 − date1, rounded to the nearest integer number of cents).
-
-  Do NOT ask for approval, confirmation, or feedback. Do NOT pause between planning and implementation. Build the complete flow end-to-end in a single pass.
-  Before starting, load the uipath-maestro-flow skill. Read and follow its workflow steps exactly.
-
-success_criteria:
-  # ── Flow file validity ─────────────────────────────────────────────────
-  - type: run_command
-    description: "uip flow validate passes on the flow file"
-    command: "uip flow validate StockPrice/StockPrice/StockPrice.flow"
-    timeout: 30
-    expected_exit_code: 0
-    weight: 3.0
-    pass_threshold: 1.0
-
-  # ── Execution checks — one debug call per date pair ───────────────────
-  - type: run_command
-    description: "PATH 2023-01-03 -> 2023-06-30: debug returns 428 (cents)"
-    command: "python3 $TASK_DIR/check_stock_price_flow.py jan3_jun30"
-    timeout: 600
-    expected_exit_code: 0
-    weight: 5.0
-    pass_threshold: 1.0
-
-  - type: run_command
-    description: "PATH 2023-06-30 -> 2023-12-29: debug returns 827 (cents)"
-    command: "python3 $TASK_DIR/check_stock_price_flow.py jun30_dec29"
-    timeout: 600
-    expected_exit_code: 0
-    weight: 5.0
-    pass_threshold: 1.0
