+++ tests/tasks/uipath-maestro-flow/wiki_pageviews/wiki_pageviews.yaml
@@ -1,0 +1,56 @@
+task_id: skill-flow-wiki-pageviews
+description: >
+  Create a UiPath Flow that takes a Wikipedia article title and two dates,
+  fetches the daily pageview count for each, and returns the difference.
+  Exercises managed HTTP (manual auth) called twice in the same flow and
+  deterministic scoring via historical pageview counts.
+tags: [uipath-maestro-flow, e2e, generate, ootb, http]
+max_iterations: 1
+
+agent:
+  type: claude-code
+  turn_timeout: 1200
+
+sandbox:
+  driver: tempdir
+  python: {}
+
+initial_prompt: |
+  Create a UiPath Flow project named "WikiPageviews" that takes `article`
+  (string) and `date1`, `date2` (strings in `YYYYMMDD` format) as inputs. Use
+  the Wikimedia pageviews API
+  (`https://wikimedia.org/api/rest_v1/metrics/pageviews/per-article/en.wikipedia/all-access/user/{article}/daily/{date}/{date}`)
+  to fetch the daily view count for each date and return the difference
+  (date2 − date1) as an integer — or, when the article does not exist and
+  the API returns no data, return the literal string `Article not found`
+  instead.
+
+  Do NOT ask for approval, confirmation, or feedback. Do NOT pause between planning and implementation. Build the complete flow end-to-end in a single pass.
+  Before starting, load the uipath-maestro-flow skill. Read and follow its workflow steps exactly.
+
+success_criteria:
+  # ── Flow file validity ─────────────────────────────────────────────────
+  - type: run_command
+    description: "uip flow validate passes on the flow file"
+    command: "uip flow validate WikiPageviews/WikiPageviews/WikiPageviews.flow"
+    timeout: 30
+    expected_exit_code: 0
+    weight: 3.0
+    pass_threshold: 1.0
+
+  # ── Execution checks — success path + error path ─────────────────────
+  - type: run_command
+    description: "UiPath 2024-01-01 -> 2024-01-15: debug returns 419 (view diff)"
+    command: "python3 $TASK_DIR/check_wiki_pageviews_flow.py uipath_success"
+    timeout: 600
+    expected_exit_code: 0
+    weight: 5.0
+    pass_threshold: 1.0
+
+  - type: run_command
+    description: "Invalid article: debug returns the string 'Article not found'"
+    command: "python3 $TASK_DIR/check_wiki_pageviews_flow.py invalid_error"
+    timeout: 600
+    expected_exit_code: 0
+    weight: 5.0
+    pass_threshold: 1.0
